dao corrected

+++ src/main/java/io/khasang/orgmanager/Dao/GenericDao.java
@@ -3,7 +3,7 @@ package io.khasang.orgmanager.dao;
 import java.util.List;
 
 public interface GenericDao<T> {
-    T get(Long id);
+    T get(Integer id);
     List<T> getAll();
     void save(T object);
     void delete(T object);
